requirements.txt
flask==2.3.3
werkzeug==2.3.7
gunicorn==21.2.0
together==0.1.7
python-dotenv==1.0.0
flask-cors==4.0.0